Lesson_4/requirements.txt
selenium
webdriver-manager
pytest
pyyaml
requests